+++ src/main/java/kitchenpos/menu/dao/MenuDao.java
@@ -2,7 +2,6 @@ package kitchenpos.menu.dao;
 
 import kitchenpos.menu.domain.Menu;
 
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuDao extends JpaRepository<Menu, Long> {
