+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -10,6 +10,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
+import kitchenpos.product.domain.Product;
 
 @Entity
 @Table(name = "menu_product")
