+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -4,12 +4,10 @@ import java.util.ArrayList;
 import java.util.stream.Collectors;
 import kitchenpos.menu.dao.MenuDao;
 import kitchenpos.order.dao.OrderDao;
-import kitchenpos.table.dao.OrderTableDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTable;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderStatusRequest;
@@ -23,32 +21,26 @@ import java.util.List;
 public class OrderService {
     private final MenuDao menuDao;
     private final OrderDao orderDao;
-    private final OrderTableDao orderTableDao;
-
     private final OrderValidator validator;
 
     public OrderService(
             final MenuDao menuDao,
             final OrderDao orderDao,
-            final OrderTableDao orderTableDao,
             final OrderValidator validator
     ) {
         this.menuDao = menuDao;
         this.orderDao = orderDao;
-        this.orderTableDao = orderTableDao;
         this.validator = validator;
     }
 
     @Transactional
     public Order create(final OrderRequest orderRequest) {
         final List<OrderLineItemRequest> orderLineItemRequests = orderRequest.getOrderLineItems();
-        final OrderTable orderTable = getOrderTable(orderRequest.getOrderTableId());
-
-        validator.validateCreate(orderLineItemRequests, orderTable);
+        validator.validateCreate(orderLineItemRequests, orderRequest.getOrderTableId());
 
         List<OrderLineItem> orderLineItems = getOrderLineItems(orderLineItemRequests);
 
-        return orderDao.save(new Order(orderTable, orderLineItems));
+        return orderDao.save(new Order(orderRequest.getOrderTableId(), orderLineItems));
     }
 
     private List<OrderLineItem> getOrderLineItems(List<OrderLineItemRequest> orderLineItemRequests) {
@@ -90,11 +82,6 @@ public class OrderService {
         return menus.stream()
                 .filter(menu -> menu.getId().equals(menuId))
                 .findFirst()
-                .orElseThrow(IllegalArgumentException::new);
-    }
-
-    private OrderTable getOrderTable(Long orderTableId) {
-        return orderTableDao.findById(orderTableId)
                 .orElseThrow(IllegalArgumentException::new);
     }
 }
