+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.application;
+package kitchenpos.product.application;
 
-import kitchenpos.menu.dao.ProductDao;
-import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.dto.ProductRequest;
+import kitchenpos.product.dao.ProductDao;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
