+++ src/main/java/kitchenpos/table/validator/TableValidator.java
@@ -1,12 +1,19 @@
 package kitchenpos.table.validator;
 
 import java.util.List;
+import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Component;
 
 @Component
 public class TableValidator {
+
+    private final OrderDao orderDao;
+
+    public TableValidator(OrderDao orderDao) {
+        this.orderDao = orderDao;
+    }
 
     public void validateChangeNumberOfGuests(OrderTable table, int numberOfGuests){
         if (numberOfGuests < 0) {
@@ -19,16 +26,13 @@ public class TableValidator {
 
     }
 
-    public void validateChangeEmpty(OrderTable table, List<Order> orders){
-        validateDinning(orders);
-
-        if (table.isGrouping()) {
+    public void validateChangeEmpty(OrderTable table){
+        List<Order> orders = orderDao.findAllByOrderTableId(table.getId());
+        if(hasDinningOrder(orders)){
             throw new IllegalArgumentException();
         }
-    }
 
-    public void validateDinning(List<Order> orders){
-        if(hasDinningOrder(orders)){
+        if (table.isGrouping()) {
             throw new IllegalArgumentException();
         }
     }
