+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.table.dto.OrderTableResponse;
