+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -5,10 +5,10 @@ import static org.assertj.core.api.Assertions.*;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.dao.ProductDao;
-import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.dto.ProductRequest;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.dao.ProductDao;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
