+++ src/test/java/kitchenpos/menu/dao/ProductDaoTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.menu.dao.ProductDao;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.dao.ProductDao;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
