+++ src/test/java/kitchenpos/menu/domain/ProductTest.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.domain;
 import static org.assertj.core.api.Assertions.*;
 
 import java.math.BigDecimal;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
